kharon/requirements.txt
psycopg2
cbor2
paho-mqtt
asyncio